python/pubsub/requirements.txt
google-cloud-pubsub==2.21.1
avro==1.11.3
protobuf===4.24.4; python_version == '3.7'
protobuf==4.25.2; python_version >= '3.8'
avro==1.11.3